Currently have DGQ 0 element for pressure hard coded. This needs to be dependent on degree. There is definitely something wrong in the interface nodes between the two domains. This needs looked into (make sure they are the same size for f_a and n_a, etc...

+++ Navier/parameters.h
@@ -90,7 +90,7 @@ namespace Parameters
 	  // FE_Q Degrees
 	  prm.declare_entry("fluid velocity degree", "2", Patterns::Integer(1),
 			  "order of the finite element to use for the fluid velocity.");
-	  prm.declare_entry("fluid pressure degree", "1", Patterns::Integer(1),
+	  prm.declare_entry("fluid pressure degree", "1", Patterns::Integer(0),
 			  "order of the finite element to use for the fluid pressure.");
 	  prm.declare_entry("structure degree", "2", Patterns::Integer(1),
 			  "order of the finite element to use for the structure displacement and velocity.");
